Remove hashring.rs
